removed state from App.js again. Nested OwnerDashboard and SitterDashboard in Dashboard. Nested Dashboard and Login in Home, with a ternary checking for a logged in user
