+++ src/components/Dashboard.js
@@ -1,41 +1,31 @@
 import React, { useState } from 'react'
-import { Card, Button, Alert } from 'react-bootstrap'
+import { Button, Card, Alert, Jumbotron, Tabs, Tab } from 'react-bootstrap';
 import { useAuth } from '../contexts/AuthContext'
 import { useHistory, Link } from 'react-router-dom'
 
-export default function Dashboard() {
-    const [error, setError] = useState('')
-    const { currentUser, logout } = useAuth()
-    const history = useHistory()
+import OwnerDashboard from './OwnerDashboard';
+import SitterDashboard from './SitterDashboard';
 
-    async function handleLogout() {
-        setError('')
-
-        try {
-            await logout()
-            history.push('/')
-        } catch {
-            setError('Failed to log out')
-        }
-    }
+const Dashboard = ({baseURL}) => {
+    const { currentUser } = useAuth();
 
     return (
-        <>
-            <Card>
-                <Card.Body>
-                    <h2 className='text-center mb-4'>Profile</h2>
-                    {error && <Alert variant='danger'>{error}</Alert>}
-                    <strong>Email:</strong> {currentUser.email}
-                    <Link to='/update-profile' className='btn btn-primary w-100 mt-3'>
-                        Update Profile
-                    </Link>
-                </Card.Body>
-            </Card>
-            <div className='w-100 text-center mt-2'>
-                <Button variant='link' onClick={handleLogout}>
-                    Log Out
-                </Button>
-            </div>
-        </>
+        <div className='dashboard'>
+            {/* //include some ternary to render OwnerDashboard or SitterDashboard based on logged in user */}
+            <Jumbotron variant='plant'>
+                <h1>Welcome back!</h1>
+                <p>This is your Plant Owner Dashboard. See below for a list of all available Sitters.</p>
+            </Jumbotron>
+            <Tabs>
+            <Tab eventKey='ownerDashboard' title='Owner Dashboard'>
+                <OwnerDashboard baseURL={baseURL}  />
+            </Tab>
+            <Tab eventKey='sitterDashboard' title='SitterDashboard'>
+                <SitterDashboard baseURL={baseURL}  />
+            </Tab>
+            </Tabs>
+        </div>
     )
 }
+
+export default Dashboard;
