+++ src/components/Home.js
@@ -1,9 +1,18 @@
-import React from 'react'
+import React from 'react';
+import { useAuth } from '../contexts/AuthContext';
+import Dashboard from './Dashboard'
+import Login from './Login'
 
-export default function Home() {
+const Home = ({baseURL}) => {
+  const { currentUser } = useAuth();
     return (
         <div>
-            
+            { currentUser ? 
+                <Dashboard baseURL={baseURL} /> :
+                <Login />
+            }
         </div>
     )
 }
+
+export default Home;
